Refactor FindProfileMessageWithAddress
Inheriting from FindProfileMessage was messy, so this commit removes it.
To enable that, it also moves methods related to building out
the core XML MPI query to FindProfileMessageHelpers. It does this to
the many methods in FindProfileMessage, as well. This makes
FindProfileMessageHelpers a better consolidation of the shared methods
used to building out the XML.