API-18932: Add ability to query MPI with addresses
Currently, MPI::Service.new.find_profile requires the given_names,
last_name, birth_date, and ssn of an individual. This commit creates an
alternative way where address info is substituted for the ssn. With
accurate address data in the query to match what's stored
in MPI, this should be enough information to yield a match in MPI.